apps/file-q-and-a/nextjs-with-flask-server/app/requirements.txt
fastapi>=0.97.0
openai>=0.27.2
langchain>=0.0.198
llama-index>=0.6.26
pinecone-client>=2.0.13
PyPDF2>=2.10.4
pypdf>=3.7.1
numpy>=1.23.2
scikit-learn>=1.1.2
docx2txt>=0.8
jinja2>=3.0.1
PyYAML>=6.0
tiktoken>=0.1.2kill
unstructured>=0.7.4
nest_asyncio>=1.5.6
rake_nltk>=1.0.6
nltk>=3.5
transformers>=4.30.2
mysql-connector-python>=8.0.33
uvicorn>=0.22.0
python-multipart==0.0.5
sqlalchemy>=2.0.16
sqlalchemy_pagination>=0.0.2
Pymysql>=1.0.3
aioboto3>=11.2.0
fastapi_jwt_auth>=0.5.0